requirements.txt
google-genai==1.52.0
pandas==2.3.3
scikit-learn==1.7.2
streamlit==1.51.0
tabulate==0.9.0
fpdf2==2.7.9
